Restore couple of tests, remove legacy disclosed proof
